requirements.txt
google-generativeai
openai
termcolor
httpx